+++ src/source.extension.cs
@@ -11,7 +11,7 @@ namespace FileDiffer
         public const string Name = "File Differ";
         public const string Description = "The easiest way to diff two files directly in solution explorer";
         public const string Language = "en-US";
-        public const string Version = "1.1";
+        public const string Version = "1.2";
         public const string Author = "Mads Kristensen";
         public const string Tags = "compare, files, diff";
     }